Move metaalmanac into the metaload module
