+++ anise/src/structure/planetocentric/ellipsoid.rs
@@ -10,6 +10,7 @@
 use core::f64::EPSILON;
 use core::fmt;
 use der::{Decode, Encode, Reader, Writer};
+use serde_derive::{Deserialize, Serialize};
 
 #[cfg(feature = "python")]
 use pyo3::exceptions::PyTypeError;
@@ -28,7 +29,7 @@ use pyo3::pyclass::CompareOp;
 /// Example: Radii of the Earth.
 ///
 ///    BODY399_RADII     = ( 6378.1366   6378.1366   6356.7519 )
-#[derive(Copy, Clone, Debug, PartialEq)]
+#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[cfg_attr(feature = "python", pyclass)]
 #[cfg_attr(feature = "python", pyo3(get_all, set_all))]
 #[cfg_attr(feature = "python", pyo3(module = "anise.astro"))]
